Add BARAS for testing nginx config logic
Added a few minimal BARAS to exercise some of the logic in our
NGINX config around the upload endpoints

For sake of test speed, these tests are designed to be very
minimal and set up the bare minimum state (i.e. no pushing apps)
required to hit these error cases. If the NGINX config logic is
removed the endpoint will still fail with this setup, but with different
error codes and messages.

[finishes #163158930]