V2/V3 app revisions mix now expects a revision to be created
We discovered a strange bug around `cf stop` not working as expected
on the second push of an app when the health check is not specified that
would cause this test to fail often. As a workaround for this test we
are now specifying a healthcheck type.

More information on that bug:
https://www.pivotaltracker.com/story/show/165516315

[#165366785]

Co-authored-by: Tim Downey <tdowney@pivotal.io>
Co-authored-by: Seth Boyles <sboyles@pivotal.io>