Specify buildpack when pushing apps
By not specifying the buildpack on these pushes, we were allowing the
staging process to download all the buildpacks in order to check for the
one to use. However, when these tests were running in parallel with
CATs, CATs was generating and deleting a bunch of test buildpacks. It
seemed like the staging process was seeing CATs-generated buildpacks,
trying to download them, and then failing when they couldn't be
downloaded (probably because CATs had already cleaned them up).

We know which buildpack these apps need, so we just specify it with the `-b`
flag in these tests.

Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>