Refresh auth token before use to reduce flakes
We modified FetchRecentLogs, UploadPackage and makeApp to no longer
accept a token as an argument. This guarantees that the token will be
valid at the time of the call.

This was causing flakes where we would instantiate a single token in the
beginning of a test and then use it throughout the test. This meant the
tests flaked if the test took longer than the tokens life.

Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Chris Selzo <cselzo@pivotal.io>