Mark the Quotas test as pending
This test alters the quota for the org that all bara tests use. This
leads to failures in other tests when running in parallel.

Eventually we should move this test into CCNG or find a way to have it
always run in serial (e.g. when Ginkgo 2.0 adds that feature).

Authored-by: Matt Royal <mroyal@pivotal.io>