Account for removed `/v3/apps/:guid/actions/apply_manifest` endpoint
We are also pending the manifest "fails the job and does not change the
memory" test as it seems like the error case is different when using the
space apply manifest endpoint. As this was an experimental endpoint to
begin with, pending this test should not impact customer facing
functionality.

[#175754135]

Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>