Unpend manifest BARA
This BARA was pended when we switched from using the app manifest
endpoint (now removed) to the space manifest endpoint. Previously, there
was an error case that was not handled properly in the space manifest
endpoint, so when the switch was made this test was pended. The error
handling behavior has now been ported over as of [this
commit](https://github.com/cloudfoundry/cloud_controller_ng/commit/6edaa2858fb04904e0521b91ed1dc8cac64341a4),
so we can now unpend this BARA.

[#175822881](https://www.pivotaltracker.com/story/show/175822881)

Authored-by: Belinda Liu <bliu@pivotal.io>