Allow multi-port-app to be compatible with cf-for-k8s
The multi-port-app relies on the `GOPACKAGENAME` environment variable to
set the name of the compiled binary. It is also required to use the go
buildpack in the TAS for VMs world. The go buildpack in the k8s
ecosystem does not respect this variable, and as a result the app would
fail to stage and BARAs using this app would break.
After investigating, it seems that the k8s go buildpack will name the
compiled binary `workspace`, so we are editing the manifest and the
start command for the multi-port-app to use this name.

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Teal Stannard <tstannard@pivotal.io>